+++ lib/constructs/s3-stack/aws-serverless-s3-stack.ts
